requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3; python_version >= '3'
discord-components==2.1.2
discord.py==1.7.3
httplib2==0.19.1
idna==3.2
multidict==5.1.0
oauthlib==3.1.1
packaging==21.0
pillow==8.4.0
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.7.2; python_version >= '3.6'
six==1.16.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
yarl==1.6.3
